python-requirements.txt
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
google-api-python-client==1.7.9
google-auth==2.13.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.7.0
gspread==5.6.2
httplib2==0.20.4
idna==3.4
oauthlib==3.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-frontmatter==1.0.0
PyYAML==5.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
typing==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.12
